src/requirements.txt
numpy==1.24.3
paddlepaddle==3.0.0rc1
paddleocr==2.10.0
opencv-python==4.10.0
#opencv-contrib-python
shapely
scikit-image
pyclipper
lmdb
tqdm
rapidfuzz
cython
Pillow
pyyaml
requests
albumentations
# to be compatible with albumentations
albucore
packaging
